Se modifican tipo de dato de IDs de uint32 a int64 para preservar consistencia con tipo de dato bigint en PostgreSQL. Se agrega CRUD para nueva entidad OperationType.

+++ Backend/Data/bonus.go
@@ -14,7 +14,7 @@ correspondiente marca.
 package data
 
 type Bonus struct {
-	Id        uint32
+	Id        int64
 	Brand     string
 	Remaining uint8
 	Amount    int32
