+++ Backend/Data/repair.go
@@ -19,10 +19,10 @@ package data
 import "time"
 
 type Repair struct {
-	Id              uint32
+	Id              int64
 	DateOfAdmission time.Time
 	DateOfRelease   time.Time
 	DateOfPickUp    time.Time
-	Id_receipt      uint32
-	Id_vehicle      uint32
+	Id_receipt      int64
+	Id_vehicle      int64
 }
