+++ Backend/Internal/Operation/operation_service.go
@@ -10,16 +10,16 @@ type Operation = data.Operation
 
 type IOperationRepository interface {
 	SaveOperation(Operation) (*Operation, error)
-	GetOperationById(uint32) (*Operation, error)
+	GetOperationById(int64) (*Operation, error)
 	GetOperationVehicleMotorType(Operation) (string, error)
 	GetAllOperations() ([]Operation, error)
-	GetAllOperationsByRepair(uint32) ([]Operation, error)
+	GetAllOperationsByRepair(int64) ([]Operation, error)
 	UpdateOperation(Operation) error
-	DeleteOperationById(uint32) error
+	DeleteOperationById(int64) error
 }
 
 type ITariffService interface {
-	GetOperationTariffValue(string, uint32) (int32, error)
+	GetOperationTariffValue(string, int64) (int32, error)
 }
 
 type Service struct {
@@ -42,7 +42,7 @@ func (serv *Service) SaveOperation(op Operation) (*Operation, error) {
 	return serv.Repository.SaveOperation(op)
 }
 
-func (serv *Service) GetOperationById(id uint32) (*Operation, error) {
+func (serv *Service) GetOperationById(id int64) (*Operation, error) {
 	return serv.Repository.GetOperationById(id)
 }
 
@@ -50,7 +50,7 @@ func (serv *Service) GetAllOperations() ([]Operation, error) {
 	return serv.Repository.GetAllOperations()
 }
 
-func (serv *Service) GetAllOperationsByRepair(id_repair uint32) ([]Operation, error) {
+func (serv *Service) GetAllOperationsByRepair(id_repair int64) ([]Operation, error) {
 	return serv.Repository.GetAllOperationsByRepair(id_repair)
 }
 
@@ -58,7 +58,7 @@ func (serv *Service) UpdateOperation(op Operation) error {
 	return serv.Repository.UpdateOperation(op)
 }
 
-func (serv *Service) DeleteOperationById(id uint32) error {
+func (serv *Service) DeleteOperationById(id int64) error {
 	return serv.Repository.DeleteOperationById(id)
 }
 
@@ -70,7 +70,7 @@ func (serv *Service) calculateBaseCost(op Operation, typeOfMotor string) (int32,
 	return int32(cost), nil
 }
 
-func (serv *Service) CalculateTotalBaseCost(id_repair uint32, typeOfMotor string) (int32, error) {
+func (serv *Service) CalculateTotalBaseCost(id_repair int64, typeOfMotor string) (int32, error) {
 	var totalCost int32 = 0
 
 	operations, err := serv.Repository.GetAllOperationsByRepair(id_repair)
