+++ Backend/Internal/Repair/repair_service.go
@@ -6,11 +6,11 @@ type Repair = data.Repair
 
 type IRepairRepository interface {
 	SaveRepair(Repair) (*Repair, error)
-	GetRepairById(uint32) (*Repair, error)
-	GetRepairByIdReceipt(uint32) (*Repair, error)
+	GetRepairById(int64) (*Repair, error)
+	GetRepairByIdReceipt(int64) (*Repair, error)
 	GetAllRepairs() ([]Repair, error)
 	UpdateRepair(Repair) error
-	DeleteRepairById(uint32) error
+	DeleteRepairById(int64) error
 }
 
 type Service struct {
@@ -21,11 +21,11 @@ func (serv *Service) SaveRepair(r Repair) (*Repair, error) {
 	return serv.Repository.SaveRepair(r)
 }
 
-func (serv *Service) GetRepairById(id uint32) (*Repair, error) {
+func (serv *Service) GetRepairById(id int64) (*Repair, error) {
 	return serv.Repository.GetRepairById(id)
 }
 
-func (serv *Service) GetRepairByIdReceipt(id_receipt uint32) (*Repair, error) {
+func (serv *Service) GetRepairByIdReceipt(id_receipt int64) (*Repair, error) {
 	return serv.Repository.GetRepairByIdReceipt(id_receipt)
 }
 
@@ -37,6 +37,6 @@ func (serv *Service) UpdateRepair(r Repair) error {
 	return serv.Repository.UpdateRepair(r)
 }
 
-func (serv *Service) DeleteRepairById(id uint32) error {
+func (serv *Service) DeleteRepairById(id int64) error {
 	return serv.Repository.DeleteRepairById(id)
 }
